[misc] update example
