+++ examples/rollout_v2/multiply_harbor_task/task.toml
@@ -1,0 +1,19 @@
+version = "1.0"
+
+[metadata]
+author_name = "Osmosis"
+difficulty = "easy"
+category = "math"
+tags = ["multiply", "rollout_v2"]
+
+[verifier]
+timeout_sec = 60.0
+
+[agent]
+timeout_sec = 120.0
+
+[environment]
+build_timeout_sec = 600.0
+cpus = 1
+memory_mb = 512
+storage_mb = 1024
