+++ examples/rollout_v2/multiply_rollout/tools.py
@@ -1,0 +1,16 @@
+from strands import tool
+
+
+@tool(name="multiply")
+def multiply_tool(a: float, b: float) -> float:
+    """
+    Multiply two numbers
+
+    # Args:
+    - a: The first number
+    - b: The second number
+
+    # Returns:
+    - The product of the two numbers
+    """
+    return a * b
